asyncio.client: Move our custom future and related functionality to future.py
--------
kirr:

We are going to enhance/improve our custom future implementation, but
before doing that let's first move corresponding code to dedicated module.
Add comment about why custom future is used.

Plain code movement and comments.

More details about why custom future is kept being used in:

- https://github.com/zopefoundation/ZEO/pull/195#issuecomment-1130240304 , and further discussion starting from
- https://github.com/zopefoundation/ZEO/pull/195#issuecomment-1135453548

Extracted from https://github.com/zopefoundation/ZEO/pull/195
